+++ package.json
@@ -19,6 +19,7 @@
     "cors": "^2.8.5",
     "docx": "^8.0.2",
     "express": "^4.18.2",
+    "express-rate-limit": "^7.0.0",
     "mammoth": "^1.6.0",
     "multer": "^1.4.5-lts.1",
     "pdf-parse": "^1.1.1",
